requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==37.0.4
dj-rest-auth==2.2.4
Django==4.0.4
django-appconf==1.0.5
django-countries==7.3.2
django-environ==0.9.0
django-imagekit==4.1.0
django-rest-knox==4.2.0
django-tinymce==3.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
pilkit==2.0
Pillow==9.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
whitenoise==6.1.0
